# apps/manager/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  kafka.BrokerOverviewItem:
    properties:
      host:
        type: string
      id:
        type: integer
      is_controller:
        type: boolean
      jvm_stats:
        $ref: '#/definitions/kafka.JVMStats'
      leader_count:
        type: integer
      network_stats:
        $ref: '#/definitions/kafka.NetworkStats'
      partition_count:
        type: integer
      port:
        type: integer
    type: object
  kafka.BrokersOverviewResponse:
    properties:
      brokers:
        items:
          $ref: '#/definitions/kafka.BrokerOverviewItem'
        type: array
      cluster:
        $ref: '#/definitions/kafka.ClusterOverviewInfo'
    type: object
  kafka.ClusterOverviewInfo:
    properties:
      broker_count:
        type: integer
      controller_id:
        type: integer
      online_brokers:
        type: integer
      online_partitions:
        type: integer
      preferred_replica_imbalance:
        type: integer
      total_partitions:
        type: integer
      under_replicated_partitions:
        type: integer
      version:
        type: string
    type: object
  kafka.CreateTopicRequest:
    properties:
      config:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      partitions:
        type: integer
      replication_factor:
        type: integer
    required:
    - name
    type: object
  kafka.JVMStats:
    properties:
      gc_count:
        type: integer
      heap_max:
        type: string
      heap_used:
        type: string
      uptime:
        type: string
    type: object
  kafka.NetworkStats:
    properties:
      bytes_in_per_sec:
        type: integer
      bytes_out_per_sec:
        type: integer
      messages_in_per_sec:
        type: integer
    type: object
  kafka.PartitionDetailInfo:
    properties:
      in_sync_replicas:
        items:
          type: integer
        type: array
      leader:
        type: integer
      offset_earliest:
        type: integer
      offset_lag:
        type: integer
      offset_latest:
        type: integer
      partition_id:
        type: integer
      replicas:
        items:
          type: integer
        type: array
      size_bytes:
        type: integer
      size_formatted:
        type: string
    type: object
  kafka.TopicDetailsResponse:
    properties:
      metrics:
        $ref: '#/definitions/kafka.TopicMetricsInfo'
      overview:
        $ref: '#/definitions/kafka.TopicOverviewInfo'
      partitions:
        items:
          $ref: '#/definitions/kafka.PartitionDetailInfo'
        type: array
    type: object
  kafka.TopicMetricsInfo:
    properties:
      bytes_per_sec:
        type: number
      messages_per_sec:
        type: number
      messages_total:
        type: integer
    type: object
  kafka.TopicOverviewInfo:
    properties:
      config:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      partition_count:
        type: integer
      replication_factor:
        type: integer
    type: object
  kafka.TopicOverviewItem:
    properties:
      bytes_per_sec:
        type: number
      is_internal:
        type: boolean
      messages_per_sec:
        type: number
      messages_total:
        type: integer
      name:
        type: string
      out_of_sync_replicas:
        type: integer
      partition_count:
        type: integer
      replication_factor:
        type: integer
      size_bytes:
        type: integer
      size_formatted:
        type: string
    type: object
  kafka.TopicsOverviewResponse:
    properties:
      summary:
        $ref: '#/definitions/kafka.TopicsSummary'
      topics:
        items:
          $ref: '#/definitions/kafka.TopicOverviewItem'
        type: array
    type: object
  kafka.TopicsSummary:
    properties:
      internal_topics:
        type: integer
      total_size_bytes:
        type: integer
      total_topics:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@sysarmor.com
    name: SysArmor Team
    url: https://sysarmor.com/support
  description: SysArmor EDR 系统的控制平面服务 API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://sysarmor.com/terms
  title: SysArmor Manager API
  version: "1.0"
paths:
...
  /services/kafka/brokers:
    get:
      consumes:
      - application/json
      description: 获取 Kafka 集群中所有 Broker 的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: Brokers 信息
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取 Kafka Brokers 信息
      tags:
      - kafka
  /services/kafka/brokers/overview:
    get:
      consumes:
      - application/json
      description: 获取 Kafka 集群中所有 Broker 的概览信息，包括集群状态、Broker 详情、性能指标等
      produces:
      - application/json
      responses:
        "200":
          description: Brokers 概览信息
          schema:
            $ref: '#/definitions/kafka.BrokersOverviewResponse'
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取 Brokers 概览信息
      tags:
      - kafka
  /services/kafka/clusters:
    get:
      consumes:
      - application/json
      description: 获取 Kafka 集群的基本信息，包括状态、Broker 数量、Topic 数量等
      produces:
      - application/json
      responses:
        "200":
          description: 集群信息
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取 Kafka 集群信息
      tags:
      - kafka
  /services/kafka/consumer-groups:
    get:
      consumes:
      - application/json
      description: 获取 Kafka 集群中所有 Consumer Group 的信息
      produces:
      - application/json
      responses:
        "200":
          description: Consumer Groups 信息
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取 Consumer Groups 信息
      tags:
      - kafka
  /services/kafka/consumer-groups/{group}:
    get:
      consumes:
      - application/json
      description: 获取指定 Consumer Group 的详细信息，包括成员、偏移量等
      parameters:
      - description: Consumer Group 名称
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Consumer Group 详细信息
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Consumer Group 不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取特定 Consumer Group 的详细信息
      tags:
      - kafka
  /services/kafka/test-connection:
    get:
      consumes:
      - application/json
      description: 测试与 Kafka 集群的连接状态
      produces:
      - application/json
      responses:
        "200":
          description: 连接成功
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 连接失败
          schema:
            additionalProperties: true
            type: object
      summary: 测试 Kafka 连接
      tags:
      - kafka
  /services/kafka/topics:
    get:
      consumes:
      - application/json
      description: 获取 Kafka 集群中所有 Topic 的详细信息列表，包含 Prometheus 指标数据，支持分页和搜索
      parameters:
      - description: 页码，默认为1
        in: query
        name: page
        type: integer
      - description: 每页数量，默认为20
        in: query
        name: limit
        type: integer
      - description: 搜索关键词
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Topics 列表（包含详细指标）
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取 Kafka Topics 列表（增强版）
      tags:
      - kafka
    post:
      consumes:
      - application/json
      description: 创建一个新的 Kafka Topic
      parameters:
      - description: 创建 Topic 请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/kafka.CreateTopicRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Topic 创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Topic 已存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 创建新的 Topic
      tags:
      - kafka
  /services/kafka/topics/{topic}:
    delete:
      consumes:
      - application/json
      description: 删除指定的 Kafka Topic
      parameters:
      - description: Topic 名称
        in: path
        name: topic
        required: true
        type: string
      - description: 强制删除，忽略错误
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Topic 删除成功
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Topic 不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 删除 Topic
      tags:
      - kafka
    get:
      consumes:
      - application/json
      description: 获取指定 Topic 的详细信息，包括分区、副本、分区大小等增强信息
      parameters:
      - description: Topic 名称
        in: path
        name: topic
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Topic 详细信息（增强版）
          schema:
            $ref: '#/definitions/kafka.TopicDetailsResponse'
        "404":
          description: Topic 不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取特定 Topic 的详细信息（增强版）
      tags:
      - kafka
  /services/kafka/topics/{topic}/config:
    get:
      consumes:
      - application/json
      description: 获取指定 Topic 的详细配置信息，包括 retention.ms、segment.ms 等所有配置参数
      parameters:
      - description: Topic 名称
        in: path
        name: topic
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Topic 配置信息
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Topic 不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取 Topic 配置
      tags:
      - kafka
    put:
      consumes:
      - application/json
      description: 更新指定 Topic 的配置参数
      parameters:
      - description: Topic 名称
        in: path
        name: topic
        required: true
        type: string
      - description: 配置更新请求
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 配置更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Topic 不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 更新 Topic 配置
      tags:
      - kafka
  /services/kafka/topics/{topic}/messages:
    get:
      consumes:
      - application/json
      description: 获取指定 Topic 的消息内容，支持分页和过滤
      parameters:
      - description: Topic 名称
        in: path
        name: topic
        required: true
        type: string
      - description: 消息数量限制，默认为10，最大100
        in: query
        name: limit
        type: integer
      - description: 指定分区
        in: query
        name: partition
        type: integer
      - description: 起始偏移量，可以是数字或'earliest'/'latest'
        in: query
        name: offset
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Topic 消息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取 Topic 消息
      tags:
      - kafka
...
  /services/kafka/topics/overview:
    get:
      consumes:
      - application/json
      description: 获取 Kafka 集群中所有 Topic 的概览信息，包括分区数量、消息数量、大小、速率等统计数据
      produces:
      - application/json
      responses:
        "200":
          description: Topics 概览信息
          schema:
            $ref: '#/definitions/kafka.TopicsOverviewResponse'
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取 Topics 概览信息
      tags:
      - kafka
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"